# openapi/media-svc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/share-items/shared-by-user": {
      "get": {
        "operationId": "ShareItemController_findItemsSharedByUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShareItemsResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-by-user/media-items": {
      "get": {
        "operationId": "ShareItemController_findMediaItemsSharedByUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-by-user/playlists": {
      "get": {
        "operationId": "ShareItemController_findPlaylistsSharedByUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-with-user": {
      "get": {
        "operationId": "ShareItemController_findItemsSharedWithUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItemsResponseDto"
                }
              }
            }
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-with-user/media-items": {
      "get": {
        "operationId": "ShareItemController_findMediaItemsSharedWithUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-with-user/playlists": {
      "get": {
        "operationId": "ShareItemController_findPlaylistsSharedWithUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/{shareId}": {
      "get": {
        "operationId": "ShareItemController_findShareItem",
        "parameters": [
          {
            "name": ":shareId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItem"
                }
              }
            }
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ShareItemController_removeShareItem",
        "parameters": [
          {
            "name": ":shareId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItem"
                }
              }
            }
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/read/{shareId}": {
      "post": {
        "operationId": "ShareItemController_readShareItem",
        "parameters": [
          {
            "name": ":shareId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItem"
                }
              }
            }
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/unshare-all-items": {
      "post": {
        "operationId": "ShareItemController_removeAllShareItems",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareItemsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/unshare-all-by-user-id": {
      "post": {
        "operationId": "ShareItemController_removeShareItemAllByUserId",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareItemsByUserIdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["share-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/playlists": {
      "post": {
        "operationId": "PlaylistController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": ["playlists"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PlaylistController_findAll",
        "parameters": [
          {
            "name": "text",
            "required": false,
            "in": "query",
            "allowEmptyValue": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "explode": true,
            "allowEmptyValue": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["playlists"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/playlists/{playlistId}": {
      "put": {
        "operationId": "PlaylistController_update",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              }
            }
          }
        },
        "tags": ["playlists"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PlaylistController_remove",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["playlists"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PlaylistController_findOne",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": ["playlists"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/playlists/{playlistId}/share/{userId}": {
      "post": {
        "operationId": "PlaylistController_share",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShareItem"
                  }
                }
              }
            }
          }
        },
        "tags": ["playlists"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Media Service",
    "description": "Media Service",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "local"
    },
    {
      "url": "https://mediashare-api-staging.herokuapp.com",
      "description": "staging"
    },
    {
      "url": "https://mediashare-api-prod.herokuapp.com",
      "description": "production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "TagKeyValue": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": ["key", "value"]
      },
...
      "ShareItem": {
        "type": "object",
        "properties": {}
      },
      "MediaItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "JWuZMCJwXFLK53lVTQ8CJiC7IWxujnylov4h9AwYvu2ar8O2a0TlJiERkTMKGcx3q4hsBsaoqVHMmrFLs25NmF7YGG51ZOZ9VTTeeMJsJkPxVvFH03CMohFcaz2bp",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "key": {
            "type": "string",
            "example": "PEnT29xwgPHwqcinagxjtSebGHJFv8fF7xpmTmt9d9T2wxyXatOYhzedLZsPRCWeFnVNiqdIQG86R6AoMr1PbShgC4j1qTYqvitek6ivy3fe9wTPqUzKVOHjveeVw",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "25fIJNU6L1BQ6NgHm5phRCi1rnZxwMdnZryAyp3DXmw2qBQWSFw6axXx9cdJ7DYPPJvkAgQIBTshQ2qtEqkNYa6vQL8v1xwSUMvJl8qmo25MXoDG0aZBPauZpdChw",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "eNfFpVtqu7rVNRxHkbQoMBofYsYkh26gDtrg7Q1GWNGcfWsx5cItamMONyVvLt0PwuTgFJXqKcMEhzAMRErxd8FY0mZoaqRvoZfVBfa6S7Ow9JHAxGKS807etOY9u",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "U3mzMkkeFpp7MPAGshiZd3vQPqBrizidERvhAEEYqJziWtK4Odz9TWFUcjHxCDzB2AhYN0tsaYnwoRQH60FmHVJ5RS5XpmuX7Eb9c9qYPMrMbkmcjVg6ko3ekrmpfCBku40zKjSEPjc3CMqz3fqgWPwYXTtp7QWwRbO0AnScoowZuAIpCc1rx9EOYp8SU29AWiTBn8QXWd9NKZOCKO81Implx4SaNU1go7RnDWLGMpTNWLVvTcbpWtFrt1zqTDGnEuQZUCnUwxvDnXxTlf2zyK14m7kqTS6QzsqX262ibVZi"
          },
          "description": {
            "type": "string",
            "example": "azUbo2T9JfKhFbR3kXOeNyE7xhaAlXLV5WUlax3jH6NxgVZv0GR65WnWhubmqSa9PtIoEiE3BCdittGlygqPbeA6pCMr3KGCXOsYuVdVY2mzjku09MlKj7nQHEESAVTrHHZNtyEUSft8VnWm9MkbLoQKD7svAA4i5yFktkkjguYIo1VbWZQp8bxClpwGS43clXCrQY7FUXwKAUcy5H9wQzULMu1q5c8rQnf4KctF8nM20dznqJWVdDJQ4jMtm7f0dBltGxWBDD1dP66tV8JYTOJL0paVNNiO95AbUWC6FYxY"
          },
          "uri": {
            "type": "string",
            "example": "7HuRvCG53qk2PWjS0FHvlpEJqFVxfbKlNvbb3NvfkNiPfvSMAOHcx2PkXXr7IPPRyMByYSN0bRXXNJBl8B2pbWa4PKu4iVuCJmsvwIB1JcBcfkvrrlGV5Q9QdFosYvfM9MuA8UMNsZwyVQ6NKKyPP3oJ7oH2iIhK6ekOKA9g019OYY9R0wytw9mQpC8rhIxTyA0FGxOPjrHGiGFpEODjavqhLPegCEtXu7aMo2Np1hjRBo1nW9qsrewox0WtaB4KnmqpHqzTd12qP30lW56jA4YD0EkpgOl3Fx24rOtc4DL5"
          },
          "imageSrc": {
            "type": "string",
            "example": "4TSbWybuQ3lqQ9TIYNvUOuNPdxtN5CnQURhHojOFCOMQAmbhqGPF4Ea4xqsxo0j5TbJOPUCi4mq3tmWytMXYdugnrgFatGxpuf0AcJI11IzvVGwUBToBZefR8Tm3Ic2R9xUXUzPfj45WAW3SS95Ft3g1MCl3CHraJ8TczbJJcjHIYFoBsxnD98I441Gabvd17oXmYs7tFfXHD0QxQNDS3EGRsAuEio1FQONCZxgoAXtvfdk3uZXxhMoMUfdYYhPNAB64dvFOQZDcYx0kvRGRvhodswEJBGKLoeFcn8VgwvmN"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "type": "string",
            "enum": ["private", "shared", "subscription", "public"]
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "eTag": {
            "type": "string",
            "example": "4iQkgfeFMUNDEYkRZHNhNo22jrrEM9mrf9XcVl5Xi0Funw34dMmYgF10Iu9K9dSNQXmPjWM7sMiA5yx480SNZGRx8kCOmjlbAEsR9M9MKn8pYBup91MdbH7bjdEXY",
            "maxLength": 255,
            "minLength": 0
          },
          "shareCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "key",
          "userId",
          "title",
          "description",
          "uri",
          "visibility",
          "shareCount",
          "viewCount",
          "likesCount"
        ]
      },
...
      "PlaylistVisibilityType": {
        "type": "string",
        "enum": ["private", "shared", "subscription", "public"]
      },
...
      "ShareItemsResponseDto": {
        "type": "object",
        "properties": {
          "mediaItems": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaShareItemResponseDto"
              }
            ]
          },
          "playlists": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PlaylistShareItemResponseDto"
              }
            ]
          }
        },
        "required": ["mediaItems", "playlists"]
      },
      "PlaylistDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "JWuZMCJwXFLK53lVTQ8CJiC7IWxujnylov4h9AwYvu2ar8O2a0TlJiERkTMKGcx3q4hsBsaoqVHMmrFLs25NmF7YGG51ZOZ9VTTeeMJsJkPxVvFH03CMohFcaz2bp",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "cloneOf": {
            "type": "string",
            "example": "6581ac478b6cdde772509888",
            "readOnly": false
          },
          "title": {
            "type": "string",
            "example": "A1ovMky7K5UGGRzyPQDVZpgczdtIq25gJ9JpJjUqMHLEp0eInYoGmOOLSeUdKEb7eDAaz9HrzmL4sESGLcroxgdjKSkvLlWgj9YHdiNvBAXCfOINwN8JBFXrop6dg",
            "maxLength": 255,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "example": "DWvNYAmokeIx04UjD3IqwQSRqRJeLS3YUkmf6mHSdd87VL7oIjPQ6Yqy43SbUYcS7snOzcHxQH0AL7xBy3XGCEsMGYFuSIu0vdzwuxwtF2uOXjS74SfF9Zh6OFBmxBmV1b4Tl2I19XI9gXS2v1KLO0icCHFXmFOE5TiPUPM46EQpbQPOcxu9NDomC0rqkHcfIb2XHQ2qxaYU9vg2jZK5sMbveNGrTqkOiDKAfv7oMHYRyxeVjwhrPl6BRWLcGyfUV2s2d8wZao8KTQ3SjIdzqZLfJ3vzlXKjkqJNOTi4N0WR"
          },
          "imageSrc": {
            "type": "string",
            "example": "fGwlKL8Qwtyyot0dXpymSw17y3BrszFDSHiruvBFM73BQIITAfSRZ4znIaJ2Gvn7QwpjWm6SSGNNR1DkzpoaqpEbgYrDVGy6VV6RcvM0BDsA94CuGOBewI0vS2P7in0SK1XPc5eotBrF4IQLD1gnJRH6oV7VeY5dSON0g0VHZo8zFWXHi1pR1DuMtTZZzwYcN0joZapSo8sLpGVjBZXyhNHnmQXVjrfEZdY0UFTGTpEG2aBMQftxr7JEUBf1w1ay0JdBfMacAhxf3fTQiN9AthVtOh0urLuB4kWDz73pTm7k"
          },
          "visibility": {
            "$ref": "#/components/schemas/PlaylistVisibilityType"
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "mediaItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItemDto"
            }
          },
          "playlistItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistItemDto"
            }
          },
          "shareCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "title",
          "description",
          "imageSrc",
          "visibility",
          "mediaItems",
          "playlistItems",
          "shareCount",
          "viewCount",
          "likesCount"
        ]
      },
      "ShareItemsDto": {
        "type": "object",
        "properties": {
          "shareItemIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["shareItemIds"]
      },
      "ShareItemsByUserIdDto": {
        "type": "object",
        "properties": {
          "shareItemByUserIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["shareItemByUserIds"]
      },
...
      "CreatePlaylistDto": {
        "type": "object",
        "properties": {
          "cloneOf": {
            "type": "string",
            "example": "6581ac478b6cdde772509888",
            "readOnly": false
          },
          "userId": {
            "type": "string",
            "example": "o2SThQmVKayyWsbUw6ySQck41dxubWIkKXEhSTGoooe5pgiZlBZIzyga3Vjti9q3gv8C68IvoM8T8eGyurv3IeNmVUHukNMAVuTK12EAjJx89DxtKrjP1x2MO1Um9",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "LYgLJAD0KjMWczi6SbPq5WOZ3xfQXFxK0a9pi2JRdaifFWMpsqdTpJxlLIkIJmnNGthoVJ3psbR8SoeKYtkQjLceMGzbQhiieZPyYMfPuIDxSIUDDi417B0ASr9hA",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "Fa8seIhuSDL6zi0X0v4hD2O9FgSmR9YfuAITm8po4lQx7DUqpz8tUkbBUc1NFBNZTOCy0gukn6PjeV5jmkHEsFDKR0XY6ODyqU4Udu9GL1oSqxdrurY2TNy83V8RVf37JQ5rBo9WJ9mk4ZuJYHKt1hTDdWhoRWal3I6TSWNs46BC26e59VU9D7S8nv3FNjKWbhPFM6g55F2QOo34weobeBLlr9CmmORYsHk1CI7c9wPLry2XvIVbyisKTKHGHNIlv2JoR3e8WQupjCoD2AyTeZ6EWwXF5ynS5a8dAUe94aB1"
          },
          "description": {
            "type": "string",
            "example": "dHLAahEHilKg05qmEwbf5RQGw5Sq8ipapa9xAW0hCCfgiZXmWNCqYDwphuEyJjSGi8kto8Znlsc91Q5QgcHm9Gh0pfMhJja38J3hs0tlLCjg3lrM0rQtxmBPzf5vb52FMCrKYfG2xjRLaU6WZKhEOEp5IkkiPaBUa7OoscV5ui9TO3SEk9hvKYJWOatQcADUmkTZGitrWxklIO1NqoyY0Aziy9Bfq0J32d64TBDARgpAu0DxUP7qYse5IQ7FtIAoFp2hK5Z4p02lU1kwu03aEIoOAegh8sdlDWWbCIpDmhIy"
          },
          "imageSrc": {
            "type": "string",
            "example": "xXCBJpxpu3fB3dF8XhLLOsy6JRIYIXO4aOgy9Xc59Gh8jrMvEG4GYuUKk25OURA6ZZIv29eE9mZTsVYn9f70aHp3OoEUw01EvfY4Xyaij3xNSfV0D2H23caf6j0i6Hf8lrMtA22tuQKW8cjU3DfDaS6ShC8XKSxcBPypkS9eqyeGYxAcrs8WbgoXzAuM5jC1b0BYYpizPg1pR6Pg6FUsxXJt4ghqts3e8Qdz5ff3lx8M7BWaMGQDkcAGN2tDb1GBSB8xly0YYZYhwCEPyFauqjA1PPibcPp55jS9AofYS8Ui"
          },
          "mediaIds": {
            "writeOnly": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "$ref": "#/components/schemas/PlaylistVisibilityType"
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          }
        },
        "required": ["userId", "title", "description", "mediaIds", "visibility"]
      },
...
      "Playlist": {
        "type": "object",
        "properties": {}
      }
    }
  }
}